requirements.txt
djangorestframework
cryptography
celery[redis]
redis
requests
websocket-client
python-binance
psycopg2-binary
